refactor: unifica patrón Vertex (auth común, fetchers/adapters delgados)
Vertex es una capa de transporte (auth GCP + endpoint + SSE), ortogonal
al provider. Cada Vertex+X ahora hereda del fetcher/adapter directo y
solo declara el delta — sin forks ni copy-paste.

Cambios:
- fetchers/vertex/_auth.py: VertexAuthMixin + JWT/OAuth/SA helpers,
  antes duplicados entre vertex_anthropic.py y vertexai.py.
- fetchers/vertex/anthropic.py y vertex/gemini.py: clientes Vertex que
  heredan de AnthropicClient/GeminiClient, ~50 líneas cada uno.
- fetchers/anthropic.py: hook _get_url(model) sobrescribible (default
  retorna BASE_URL — no-regresión bit a bit del cliente directo).
- adapters/vertex_gemini_adapter.py: paralelo a vertex_anthropic_adapter.
- adapters/gemini_adapter.py: __init__ simplificado, solo api_key.
- core.py: adapter_map["vertexai"] apunta a VertexGeminiAdapter.
- Borrados fetchers/vertex_anthropic.py (291) y vertexai.py (622).

Bugs corregidos por el refactor:
- Vertex Anthropic streaming: el stream=False hardcodeado en
  _build_request_body se elimina al heredar del padre.
- Vertex Anthropic acepta los 4 params Anthropic que antes perdía
  silenciosamente: container, context_management, mcp_servers,
  service_tier.
- Auth GCP deja de vivir en vertexai.py (acoplamiento espurio entre
  fetchers hermanos), pasa a fetchers/vertex/_auth.py.

API pública de InstantNeo intacta. .gitignore añade patrón global ".*"
con whitelist para sandbox local de credenciales.

Validado contra Vertex AI real: 7/7 smoke tests pasan (Anthropic Haiku
4.5 en us-east5, Gemini 3.1 flash-lite en global) cubriendo non-stream,
streaming SSE y function calling para los dos providers.

Doc de revisión en docs/es/refactor-vertex-unification.md.
