+++ instantneo/fetchers/__init__.py
@@ -3,13 +3,16 @@ Fetchers para proveedores de LLM.
 
 Este módulo contiene clientes HTTP puros para interactuar con diferentes
 proveedores de modelos de lenguaje, sin depender de sus SDKs oficiales.
+
+Providers directos: anthropic, openai, groq, cerebras, gemini.
+Providers en Vertex AI: ver subpaquete `vertex/`.
 """
 
 from .anthropic import fetch_anthropic
 from .openai import fetch_openai
 from .groq import fetch_groq
 from .cerebras import fetch_cerebras
-from .vertex_anthropic import VertexAnthropicClient
+from .vertex import VertexAnthropicClient, VertexGeminiClient
 
 __all__ = [
     "fetch_anthropic",
@@ -17,4 +20,5 @@ __all__ = [
     "fetch_groq",
     "fetch_cerebras",
     "VertexAnthropicClient",
+    "VertexGeminiClient",
 ]
